+++ timeseries/apps/main.py
@@ -1,0 +1,48 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+import os
+import dash
+import dash_core_components as dcc
+import dash_html_components as html
+import pandas as pd
+import plotly.graph_objs as go
+from app import app
+from dash.dependencies import Input, Output
+
+df = pd.read_csv('https://raw.githubusercontent.com/plotly/datasets/master/hello-world-stock.csv')
+df['Date'] = pd.to_datetime(df.Date, infer_datetime_format=True)
+
+if 'DYNO' in os.environ:
+    app_name = os.environ['DASH_APP_NAME']
+else:
+    app_name = 'dash-timeseriesplot'
+
+layout = html.Div([html.H1("Stock Prices", style={'textAlign': 'center'}),
+    dcc.Dropdown(id='my-dropdown',options=[{'label': 'Tesla', 'value': 'TSLA'},{'label': 'Apple', 'value': 'AAPL'},{'label': 'Coke', 'value': 'COKE'}],
+        multi=True,value=['TSLA'],style={"display": "block","margin-left": "auto","margin-right": "auto","width": "60%"}),
+    dcc.Graph(id='my-graph')
+], className="container")
+
+
+@app.callback(Output('my-graph', 'figure'),
+              [Input('my-dropdown', 'value')])
+def update_graph(selected_dropdown_value):
+    dropdown = {"TSLA": "Tesla","AAPL": "Apple","COKE": "Coke",}
+    trace1 = []
+    trace2 = []
+    for stock in selected_dropdown_value:
+        trace1.append(go.Scatter(x=df[df["Stock"] == stock]["Date"],y=df[df["Stock"] == stock]["Open"],mode='lines',
+            opacity=0.7,name=f'Open {dropdown[stock]}',textposition='bottom center'))
+        trace2.append(go.Scatter(x=df[df["Stock"] == stock]["Date"],y=df[df["Stock"] == stock]["Close"],mode='lines',
+            opacity=0.6,name=f'Close {dropdown[stock]}',textposition='bottom center'))
+    traces = [trace1, trace2]
+    data = [val for sublist in traces for val in sublist]
+    figure = {'data': data,
+        'layout': go.Layout(colorway=["#5E0DAC", '#FF4F00', '#375CB1', '#FF7400', '#FFF400', '#FF0056'],
+            height=600,title=f"Opening and Closing Prices for {', '.join(str(dropdown[i]) for i in selected_dropdown_value)} Over Time",
+            xaxis={"title":"Date",
+                   'rangeselector': {'buttons': list([{'count': 1, 'label': '1M', 'step': 'month', 'stepmode': 'backward'},
+                                                      {'count': 6, 'label': '6M', 'step': 'month', 'stepmode': 'backward'},
+                                                      {'step': 'all'}])},
+                   'rangeslider': {'visible': True}, 'type': 'date'},yaxis={"title":"Price (USD)"})}
+    return figure
