+++ timeseries/requirements.txt
@@ -1,0 +1,33 @@
+attrs==19.1.0
+certifi==2019.3.9
+chardet==3.0.4
+Click==7.0
+dash==0.42.0
+dash-core-components==0.47.0
+dash-html-components==0.16.0
+dash-renderer==0.23.0
+dash-table==3.6.0
+decorator==4.4.0
+Flask==1.0.2
+Flask-Compress==1.4.0
+gunicorn==19.9.0
+idna==2.8
+ipython-genutils==0.2.0
+itsdangerous==1.1.0
+Jinja2==2.10.1
+jsonschema==3.0.1
+jupyter-core==4.4.0
+MarkupSafe==1.1.1
+nbformat==4.4.0
+numpy==1.16.3
+pandas==0.24.2
+plotly==3.8.1
+pyrsistent==0.15.1
+python-dateutil==2.8.0
+pytz==2019.1
+requests==2.21.0
+retrying==1.3.3
+six==1.12.0
+traitlets==4.3.2
+urllib3==1.24.2
+Werkzeug==0.15.2
